WebSockets/tests/submissions/Opera/cookies/007.html

<!doctype html>
<title>WebSockets: when to process set-cookie fields in ws response</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js></script>
<div id=log></div>
<script>
setup({timeout:12100});
async_test(function(t) {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/ws/set-cookie?'+encodeURIComponent('ws_test=test; Path=/'));
  ws.onopen = t.step_func(function(e) {
    ws.close();
    assert_regexp_match(document.cookie, /ws_test=test/);
    // remove cookie
    document.cookie = 'ws_test=; Path=/; Expires=Sun, 06 Nov 1994 08:49:37 GMT';
    t.done();
  });
  ws.onclose = t.step_func(function() {assert_unreached()});

// sleep for 2 seconds with sync xhr
  var sleep = new XMLHttpRequest();
  sleep.open('GET', 'support/sleep_2.php', false);
  sleep.send(null);

  if (/ws_test=test/.test(document.cookie)) {
    assert_unreached('cookie was set during script execution');
  }
}, null, {timeout:12000});
</script>
